Compute/requirements.txt
Django==2.0.6
djangorestframework==3.8.2
gunicorn==19.8.1
mysqlclient==1.3.12
Pillow==5.1.0
pkg-resources==0.0.0
pytz==2018.4
unicorn==1.0.1
